refactor: melhorando os DTOs de local e os controllers
