+++ src/main/java/com/labweb/agrodoa_backend/dto/local/CidadeDTO.java
@@ -1,0 +1,11 @@
+package com.labweb.agrodoa_backend.dto.local;
+
+import lombok.Getter;
+import lombok.AllArgsConstructor;
+
+@Getter
+@AllArgsConstructor 
+public class CidadeDTO {
+    private String id;
+    private String nome;
+}
