+++ src/main/java/com/labweb/agrodoa_backend/dto/pessoas/usuario/UsuarioRespostaDTO.java
@@ -2,9 +2,9 @@ package com.labweb.agrodoa_backend.dto.pessoas.usuario;
 
 import java.util.ArrayList;
 
-import com.labweb.agrodoa_backend.dto.LocalDTO;
 import com.labweb.agrodoa_backend.dto.RelacaoBeneficiarioDTO;
 import com.labweb.agrodoa_backend.dto.anuncio.AnuncioRespostaDTO;
+import com.labweb.agrodoa_backend.dto.local.LocalDTO;
 import com.labweb.agrodoa_backend.model.pessoas.Usuario;
 
 import lombok.Getter;
